fix: removes walls from size of the game field
Currently walls are in the coordinates of the defined game field from file.
However they should be outside of the game field so a use who makes
a level does not have to take into account the walls.
To not have negative coordinates for the walls all the coordinates
get transposed from gamefield.size to gamefield.size + wallThickness