refactor: removes the types MaxPositionX and MaxPositionY
Types MaxPositonX/Y were introduced to prevent an object to leave the grid.
However game physics always checks to reflect the ball or put the platform
before a collision occurs so it is not a problem.
I added this types in a Version of the game which only used the
Console to display the grid. There the gamefield was stored in a array.
Leaving the grid there could result in a crash but since here the
rendering is done with SDL it cannot happen anymore.